refactor: extract `types-rs` library with some common types
