+++ .circleci/config.yml
@@ -1061,6 +1061,29 @@ jobs:
       - store_test_results:
           path: libs/types/reports/
 
+  # @votingworks/types-rs
+  test-libs-types-rs:
+    executor: nodejs
+    resource_class: xlarge
+    steps:
+      - checkout-and-install
+      - run:
+          name: Build
+          command: |
+            pnpm --dir libs/types-rs build
+      - run:
+          name: Lint
+          command: |
+            pnpm --dir libs/types-rs lint
+      - run:
+          name: Test
+          command: |
+            pnpm --dir libs/types-rs test
+          environment:
+            JEST_JUNIT_OUTPUT_DIR: ./reports/
+      - store_test_results:
+          path: libs/types-rs/reports/
+
   # @votingworks/ui
   test-libs-ui:
     executor: nodejs
@@ -1191,6 +1214,7 @@ workflows:
       - test-libs-res-to-ts
       - test-libs-test-utils
       - test-libs-types
+      - test-libs-types-rs
       - test-libs-ui
       - test-libs-usb-drive
       - test-libs-utils
