+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
-
-members = ["libs/ballot-interpreter"]
+resolver = "2"
+members = ["libs/ballot-interpreter", "libs/types-rs"]
 
 [workspace.dependencies]
 base64 = "0.21.4"
