+++ libs/ballot-interpreter/Cargo.toml
@@ -2,7 +2,7 @@
 name = "ballot-interpreter"
 version = "0.1.0"
 license = "AGPL-3.0"
-edition = "2018"
+edition = "2021"
 exclude = ["build/rust_addon.node"]
 
 [lib]
@@ -30,6 +30,7 @@ hex = { workspace = true }
 zbar-rust = { workspace = true }
 base64 = { workspace = true }
 thiserror = { workspace = true }
+types-rs = { workspace = true }
 
 [dev-dependencies]
 proptest = { workspace = true }
