+++ libs/types-rs/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "types-rs"
+version = "0.1.0"
+license = "AGPL-3.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
+
+[dev-dependencies]
+proptest = { workspace = true }
